requirements.txt
llama-index
chromadb
openai
streamlit
pypdf
python-dotenv
